+++ server-jupnp/src/main/java/apincer/music/server/jupnp/MediaServerHubImplOld.java
@@ -1116,7 +1116,7 @@ public class MediaServerHubImplOld implements MediaServerHub {
     private String createDidlLiteMetadata(Track song, String songUrl) {
         String objectClass = "object.item.audioItem.musicTrack";
         String duration = formatDurationForDidl((long) (song.getAudioDuration() * 1000.0));
-        String bitrate = String.valueOf(song.getAudioBitRate() * 1024 / 8); // bps to Bps
+        String bitrate = song.getAudioBitRate() > 0 ? String.valueOf(song.getAudioBitRate() / 8) : null;
         String sampleRate = String.valueOf(song.getAudioSampleRate());
         String bitsPerSample = String.valueOf(song.getAudioBitsDepth());
         String mimeType = MimeTypeUtils.getMimeTypeFromPath(song.getPath());